Adding a way to display logit vector for #112 and add type hinting for #830
