+++ transformer_lens/HookedEncoder.py
@@ -31,7 +31,8 @@ from transformer_lens.components import (
 from transformer_lens.components.mlps.gated_mlp import GatedMLP
 from transformer_lens.config.HookedTransformerConfig import HookedTransformerConfig
 from transformer_lens.FactoredMatrix import FactoredMatrix
-from transformer_lens.hook_points import HookedRootModule, HookPoint
+from transformer_lens.hook_points import HookPoint
+from transformer_lens.HookedRootModule import HookedRootModule
 from transformer_lens.utilities import devices
 
 T = TypeVar("T", bound="HookedEncoder")
